Refactor project structure, fix imports, and add configuration files
