+++ .env.local.example
@@ -1,0 +1,21 @@
+# Example environment configuration file
+# Copy this to .env.local and replace with your actual values
+
+# Gemini AI API Key (replace with your actual key from Google AI Studio)
+GEMINI_API_KEY=your_gemini_api_key_here
+
+# MongoDB connection string (replace with your actual MongoDB URI)
+MONGODB_URI=mongodb+srv://username:password@cluster.mongodb.net/database_name?retryWrites=true&w=majority
+
+# Backend server port
+BACKEND_PORT=5001
+
+# JWT secret for authentication (generate a strong, unique secret)
+JWT_SECRET=your_secure_jwt_secret_here_replace_with_random_string
+
+# JWT token expiration time
+JWT_EXPIRES_IN=90d
+
+# Frontend Gemini API Key (if frontend needs direct access)
+# Best practice: Frontend should call backend, which uses GEMINI_API_KEY
+VITE_GEMINI_API_KEY=your_gemini_api_key_here
