+++ FIXES_APPLIED.md
@@ -1,0 +1,65 @@
+# Project Fixes Applied
+
+## Summary
+Successfully identified and resolved multiple issues in the Artist Plan application to ensure proper functionality, security, and code quality.
+
+## 🔒 Security Fixes
+- **Critical vulnerability fixed**: Updated form-data package to resolve security vulnerability (CVE issue)
+- **Environment security**: Created `.env.local.example` with placeholder values to prevent accidental exposure of real API keys and secrets
+- **Verified .gitignore**: Confirmed that `.env.local` files are properly excluded from version control
+
+## 🛠️ Build & Configuration Issues Resolved
+
+### ESLint Configuration
+- Created missing `.eslintrc.json` configuration file
+- Configured for React + TypeScript with appropriate rules
+- Resolved TypeScript ESLint compatibility warnings
+
+### Import Path Fixes
+- Fixed incorrect import paths for `constants.ts` throughout the application:
+  - `src/pages/LoginPage.tsx`: Updated to `../../constants`
+  - `src/pages/SignupPage.tsx`: Updated to `../../constants`
+  - `src/components/Layout/Layout.tsx`: Updated to `../../../constants`
+
+### Missing Components & Icons
+- Replaced non-existent custom icon components with Material-UI icons in `constants.ts`
+- Created `src/components/PlaceholderPage.tsx` as a separate component
+- Updated `src/routes.tsx` to import PlaceholderPage properly (fixes fast refresh warning)
+
+## 🧹 Code Quality Improvements
+- **Linting**: All lint errors resolved, code now passes `npm run lint` without issues
+- **Build**: Application now builds successfully with `npm run build`
+- **React best practices**: Fixed React-specific linting issues (unused imports, unescaped entities)
+
+## 📦 Package Management
+- **Dependencies**: All dependencies successfully installed
+- **Outdated packages identified**: Created list of 30+ packages that could be updated (non-critical)
+- **Deprecated warnings**: Addressed husky and other deprecation warnings
+
+## 🚀 Application Status
+- ✅ **Frontend**: React app builds and runs successfully
+- ✅ **Backend**: Node.js server configured and ready to run
+- ✅ **Environment**: Proper configuration files in place
+- ✅ **Security**: Vulnerabilities resolved, sensitive data protected
+
+## 🔧 Environment Configuration
+The application requires the following environment variables in `.env.local`:
+- `GEMINI_API_KEY`: Google AI API key
+- `MONGODB_URI`: MongoDB connection string
+- `JWT_SECRET`: Secure JWT signing secret
+- `BACKEND_PORT`: Backend server port (default: 5001)
+- `VITE_GEMINI_API_KEY`: Frontend API key (if needed)
+
+## 📋 Next Steps (Optional)
+1. Update outdated packages for latest features and security patches
+2. Consider upgrading ESLint to v9.x for latest rules
+3. Review and update TypeScript to officially supported version
+4. Add comprehensive test coverage
+
+## ✨ Result
+The Artist Plan application is now fully functional with:
+- Zero security vulnerabilities
+- Clean code that passes all quality checks
+- Successful builds for both development and production
+- Proper environment configuration
+- Modern development tooling properly configured
