+++ src/App.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { useRoutes } from 'react-router-dom';
 import { routes } from './routes'; // Using the manually defined routes for now
