+++ src/components/PlaceholderPage.tsx
@@ -1,0 +1,14 @@
+import React from 'react';
+
+interface PlaceholderPageProps {
+  title: string;
+}
+
+const PlaceholderPage: React.FC<PlaceholderPageProps> = ({ title }) => (
+  <div style={{ padding: '20px' }}>
+    <h2>{title}</h2>
+    <p>This page is under construction.</p>
+  </div>
+);
+
+export default PlaceholderPage;
