+++ src/pages/SignupPage.tsx
@@ -14,8 +14,8 @@ import {
 import Avatar from '@mui/material/Avatar';
 import LockOutlinedIcon from '@mui/icons-material/LockOutlined'; // Or a different icon for signup
 import useAuthStore from '../store/authStore';
-import { signupUser, SignupData } from '../services/authService'; // Import SignupData
-import { APP_NAME } from '../constants';
+import { signupUser, SignupData } from '../services/authService';
+import { APP_NAME } from '../../constants';
 
 const SignupPage: React.FC = () => {
   const navigate = useNavigate();
